+++ src/GPU.java
@@ -1,5 +1,5 @@
 public class GPU extends Item{
-    public GPU (String name, Float price, int wattage){ //constructor
+    public GPU (String name, double price, int wattage){ //constructor
         this.setItemName(name);
         this.setItemCost(price);
         this.setWattage(wattage);
