+++ src/Motherboard.java
@@ -1,5 +1,5 @@
 public class Motherboard extends Item{
-    public Motherboard (String name, Float price, int wattage){ //constructor
+    public Motherboard (String name, double price, int wattage){ //constructor
         this.setItemName(name);
         this.setItemCost(price);
         this.setWattage(wattage);
